+++ docs/changelog.rst
@@ -94,3 +94,4 @@ Changes in Version 2.0.1
 * Allows suppressing of InsecureRequestWarning messages
 * Changes the exception isinstance check to look for basestring/str instead of
   bytes in order to properly store the error description.
+* Adds a new API for modifying snapshots.